Added ability to add, edit, remove permissions from front end. Prepared for i18n of constants file. Moved management of organizations into settings menu.

+++ rest_api/src/main/java/com/easyledger/api/service/PersonService.java
@@ -122,7 +122,7 @@ public class PersonService {
         			break;
         		
         		case "currentOrganizationId":
-        			person.setCurrentOrganizationId((long) v);
+        			person.setCurrentOrganizationId(Long.valueOf(((Integer) v).longValue()));
         			break;
         	}  	
         }
